Run optimized query if relationship present
In the `personById` query, run an optimised query rather than relying on
JPA + its lazy fields since we know at query time what fields were
requested by the GraphQL query.

Uses `LEFT JOIN FETCH` to improve the query and initialise lazy init
fields straight away. Tried my best but the JPQL query still did extra
queries even though all the information was included in the first query
it executes...

`PersonRelationshipsDataFetcher` checks if the `PersonDTO` that
surrounds the `relationship` field in the GraphQL query has already
initialised its `relationships` field. If it has, then it returns them,
otherwise is performs a query to get the relationship information.

This had to be done because the library always calls `DataFetcher`s that
are defined for a schema's field, even if they were already filled in by
a previous `DataFetcher`.